+++ src/engine/HasPredicateScan.h
@@ -65,7 +65,7 @@ class HasPredicateScan : public Operation {
   float getMultiplicity(size_t col) override;
 
  private:
-  size_t getSizeEstimateBeforeLimit() override;
+  uint64_t getSizeEstimateBeforeLimit() override;
 
  public:
   size_t getCostEstimate() override;
